+++ src/stage.c
@@ -1,7 +1,7 @@
 #include "stage.h"
 
 void draw_stage(const Stage *stage) {
-    for (int i = 0; i < stage->count; i++) {
+    for (size_t i = 0; i < stage->count; i++) {
         DrawRectangleRec(stage->platforms[i], RED);
     }
 }
